vs1

+++ redConexion/src/main/java/Models/ProductGranModel.java
@@ -1,13 +1,13 @@
-
 package Models;
 
 import lombok.Data;
+import java.math.BigDecimal;
 
 @Data
 public class ProductGranModel {
     private int idProductGra;
     private String nameGra;
     private int stock;
-    private float price;
-    
+    private BigDecimal price;
 }
+
